fastapi_nou/app/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose 
python-multipart 